# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tugabuy",
    "description": "This page contains the documentation for the entire project API. It explains all the existing routes, how they are used and the various possible responses. With the right credentials you can test the routes.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.tugabuy.ss-centi.com",
      "description": "Server on AWS"
    },
    {
      "url": "http://localhost",
      "description": "Server on Docker"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "66fd6e1af196280b6cd9627e"
          },
          "name": {
            "type": "string",
            "example": "Jonh Doe"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "$2a$10$EDqgf/HJjZKqc8H/E3hte1w/TFDcv1CmBPbQsRKPgE0G1GHnHzMO"
          },
          "phone": {
            "type": "number",
            "example": "999 999 999"
          },
          "image": {
            "type": "string",
            "example": "profile.png"
          },
          "address": {
            "type": "string",
            "example": "Braga"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-11-21T13:59:42.944Z"
          },
          "__v": {
            "type": "integer",
            "example": 0
          }
        }
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "46fd6e1af19628de6cd9627e"
          },
          "name": {
            "type": "string",
            "example": "Nike Air force 1 Low"
          },
          "desc": {
            "type": "string",
            "example": "Size: 42 / Not used"
          },
          "price": {
            "type": "number",
            "example": "70"
          },
          "id_category": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "state": {
            "type": "string",
            "example": "Active"
          },
          "id_user": {
            "type": "string",
            "example": "66fd6e1af196280b6cd9627e"
          },
          "image": {
            "type": "string",
            "example": "product.png"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-11-21T13:59:42.944Z"
          },
          "__v": {
            "type": "integer",
            "example": 0
          }
        }
      },
      "Chat": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64a8f3e25bde312c8c2f9bd7"
          },
          "id_product": {
            "type": "string",
            "example": "46fd6e1af19628de6cd9627e"
          },
          "id_user": {
            "type": "string",
            "example": "66fd6e1af196280b6cd9627e"
          },
          "__v": {
            "type": "integer",
            "example": 0
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get user by ID",
        "description": "This endpoint retrieves a user by their ID. It first checks the Redis cache for the user's data. If found, it returns the data from the cache. If not found in cache, it queries the MongoDB database and stores the result in Redis for future use.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the user to retrieve.",
            "schema": {
              "type": "string",
              "example": "66fd6e1af196280b6cd9627e"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found and returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "action": {
                      "type": "string",
                      "example": "Sucess"
                    },
                    "results": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access Denied, missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 403
                    },
                    "action": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "Access denied!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 404
                    },
                    "action": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "No user with id: 66fd6e1af196280b6cd9627e"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete a user by ID",
        "description": "This endpoint delete a user by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the user to retrieve.",
            "schema": {
              "type": "string",
              "example": "66fd6e1af196280b6cd9627e"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "403": {
            "description": "Access Denied, missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 403
                    },
                    "action": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "Access denied!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 404
                    },
                    "action": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "No user with id: {id}"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Update user by ID",
        "description": "This endpoint update user data by ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of user to update",
            "schema": {
              "type": "string",
              "example": "66fd6e1af196280b6cd9627e"
            }
          },
          {
            "in": "body",
            "name": "user",
            "required": true,
            "description": "User data allowed to update",
            "schema": {
              "type": "object",
              "properties": {
                "address": {
                  "type": "string",
                  "example": "newPassword123"
                },
                "email": {
                  "type": "string",
                  "example": "user@example.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated with sucess",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "action": {
                      "type": "string",
                      "example": "Sucess"
                    },
                    "results": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access Denied, missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 403
                    },
                    "action": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "Access denied!"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid update options provided",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "action": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "Invalid update options provided."
                    }
                  }
                }
              }
            }
          },"409": {
            "description": "Email or phone number already in use",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 409
                    },
                    "action": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "Email or phone number already in use."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/products/user/{userId}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Get product by User ID",
        "description": "This endpoint retrieves a product by user ID. It first checks the Redis cache for the product's data. If found, it returns the data from the cache. If not found in cache, it queries the MongoDB database and stores the result in Redis for future use.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the user to retrieve.",
            "schema": {
              "type": "string",
              "example": "66fd6e1af196280b6cd9627e"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products found and returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "action": {
                      "type": "string",
                      "example": "Sucess"
                    },
                    "results": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access Denied, missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 403
                    },
                    "action": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "Access denied!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Products not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 404
                    },
                    "action": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "No products found with user id: 66fd6e1af196280b6cd9627e"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/chats/user/{userId}": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Get chat by User ID",
        "description": "This endpoint retrieves a chat by user ID. It first checks the Redis cache for the chat's data. If found, it returns the data from the cache. If not found in cache, it queries the MongoDB database and stores the result in Redis for future use.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the user to retrieve.",
            "schema": {
              "type": "string",
              "example": "66fd6e1af196280b6cd9627e"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chats found and returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "action": {
                      "type": "string",
                      "example": "Sucess"
                    },
                    "results": {
                      "$ref": "#/components/schemas/Chat"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access Denied, missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 403
                    },
                    "action": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "Access denied!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Chats not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "httpCode": {
                      "type": "integer",
                      "example": 404
                    },
                    "action": {
                      "type": "string",
                      "example": "Failed"
                    },
                    "error": {
                      "type": "string",
                      "example": "No chats found with user id: 66fd6e1af196280b6cd9627e"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}